+++ ethers-core/src/types/transaction/response.rs
@@ -113,6 +113,11 @@ pub struct Transaction {
 
     #[serde(rename = "chainId", default, skip_serializing_if = "Option::is_none")]
     pub chain_id: Option<U256>,
+
+    /// Captures unknown fields such as additional fields used by L2s
+    #[cfg(not(feature = "celo"))]
+    #[serde(flatten)]
+    pub other: crate::types::OtherFields,
 }
 
 impl Transaction {
@@ -550,6 +555,7 @@ mod tests {
                 16,
             )
             .unwrap(),
+            other: Default::default(),
         };
         println!("0x{}", hex::encode(&tx.rlp()));
         assert_eq!(
@@ -593,6 +599,7 @@ mod tests {
                 16,
             )
             .unwrap(),
+            other: Default::default(),
         };
         println!("0x{}", hex::encode(&tx.rlp()));
         assert_eq!(
@@ -626,7 +633,8 @@ mod tests {
             chain_id: Some(U256::from(1)),
             access_list: None,
             max_fee_per_gas: None,
-            max_priority_fee_per_gas: None
+            max_priority_fee_per_gas: None,
+            other: Default::default()
         };
         assert_eq!(
             tx.rlp(),
@@ -672,6 +680,7 @@ mod tests {
             max_priority_fee_per_gas: Some(1500000000.into()),
             max_fee_per_gas: Some(1500000009.into()),
             chain_id: Some(5.into()),
+            other: Default::default(),
         };
         assert_eq!(
             tx.rlp(),
@@ -717,6 +726,7 @@ mod tests {
             max_priority_fee_per_gas: Some(1500000000.into()),
             max_fee_per_gas: Some(1500000009.into()),
             chain_id: Some(5.into()),
+            other: Default::default(),
         };
 
         let rlp_bytes = hex::decode("02f86f05418459682f008459682f098301a0cf9411d7c2ab0d4aa26b7d8502f6a7ef6844908495c28084e5225381c001a01a8d7bef47f6155cbdf13d57107fc577fd52880fa2862b1a50d47641f8839419a03279bbf73fde76de83440d04b9d97f3809fec8617d3557ee40ac3e0edc391514").unwrap();
@@ -762,6 +772,7 @@ mod tests {
             max_priority_fee_per_gas: Some(1500000000.into()),
             max_fee_per_gas: Some(1500000009.into()),
             chain_id: Some(5.into()),
+            other: Default::default(),
         };
 
         assert_eq!(tx.hash, tx.hash());